requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.0
pandas>=2.0.0
Pillow>=10.0.0
imagehash>=4.3.0
pytesseract>=0.3.10
playwright>=1.40.0
